# docker_images/csharp/wrapper/src/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "REST API definition for End-to-end testing of the Azure IoT SDKs.  All SDK APIs that are tested by our E2E tests need to be defined in this file.  This file takes some liberties with the API definitions.  In particular, response schemas are undefined, and error responses are also undefined.",
    "version" : "1.0.0",
    "title" : "Azure IOT End-to-End Test Wrapper Rest Api"
  },
...
  "schemes" : [ "https" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/control/message" : {
      "put" : {
        "tags" : [ "control" ],
        "summary" : "log a message to output",
        "operationId" : "Control_LogMessage",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "logMessage",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/logMessage"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/control/flags" : {
      "put" : {
        "tags" : [ "control" ],
        "summary" : "set flags for the objects in this server to use",
        "operationId" : "Control_SetFlags",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "flags",
          "required" : true,
          "schema" : {
            "type" : "object"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/control/command" : {
      "put" : {
        "tags" : [ "control" ],
        "summary" : "send an arbitrary command",
        "operationId" : "Control_SendCommand",
        "parameters" : [ {
          "name" : "cmd",
          "in" : "query",
          "description" : "command string",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/service/connect" : {
      "put" : {
        "tags" : [ "service" ],
        "summary" : "Connect to service",
        "description" : "Connect to the Azure IoTHub service.  More specifically, the SDK saves the connection string that is passed in for future use.",
        "operationId" : "Service_Connect",
        "parameters" : [ {
          "name" : "connectionString",
          "in" : "query",
          "description" : "connection string",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/connectResponse"
            }
          }
        }
      }
    },
    "/service/{connectionId}/disconnect/" : {
      "put" : {
        "tags" : [ "service" ],
        "summary" : "Disconnect from the service",
        "description" : "Disconnects from the Azure IoTHub service.  More specifically, closes all connections and cleans up all resources for the active connection",
        "operationId" : "Service_Disconnect",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/service/{connectionId}/moduleMethod/{deviceId}/{moduleId}" : {
      "put" : {
        "tags" : [ "service" ],
        "summary" : "call the given method on the given module",
        "operationId" : "Service_InvokeModuleMethod",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "moduleId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "methodInvokeParameters",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/methodInvoke"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
    "/service/{connectionId}/deviceMethod/{deviceId}" : {
      "put" : {
        "tags" : [ "service" ],
        "summary" : "call the given method on the given device",
        "operationId" : "Service_InvokeDeviceMethod",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "methodInvokeParameters",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/methodInvoke"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
    "/service/{connectionId}/sendC2d/{deviceId}" : {
      "put" : {
        "tags" : [ "service" ],
        "summary" : "Send a c2d message",
        "operationId" : "Service_SendC2d",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "eventBody",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/eventBody"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "logMessage" : {
      "type" : "object",
      "properties" : {
        "message" : {
          "type" : "string",
          "description" : "text of message"
        }
      },
      "description" : "message from the test script to output to the log"
    },
    "connectResponse" : {
      "type" : "object",
      "properties" : {
        "connectionId" : {
          "type" : "string",
          "description" : "unique identifier for this connection"
        }
      },
      "description" : "result of a connection to a service, device, or module client",
      "example" : {
        "connectionId" : "connectionId"
      }
    },
...
    "eventBody" : {
      "type" : "object",
      "properties" : {
        "body" : {
          "type" : "object",
          "description" : "payload to send to the method",
          "properties" : { }
        },
        "horton_flags" : {
          "type" : "object",
          "description" : "flags used by horton",
          "properties" : { }
        },
        "attributes" : {
          "type" : "object",
          "description" : "Message attributes",
          "properties" : { }
        }
      },
      "description" : "body for an invoming or outgoing event or message",
      "example" : {
        "horton_flags" : "{}",
        "attributes" : "{}",
        "body" : "{}"
      }
    },
...
    "methodInvoke" : {
      "type" : "object",
      "properties" : {
        "methodName" : {
          "type" : "string",
          "description" : "name of method to invoke"
        },
        "payload" : {
          "type" : "object",
          "description" : "payload to send to the method",
          "properties" : { }
        },
        "responseTimeoutInSeconds" : {
          "type" : "integer",
          "description" : "number of seconds to wait for a response"
        },
        "connectTimeoutInSeconds" : {
          "type" : "integer",
          "description" : "number of seconds to wait for the module to connect"
        }
      },
      "description" : "parameters used to invoke a method"
    }
  }
}